Initial commit for client_portal-assets
This repo is intended to hold static content (e.g. images, colours, etc)
which will be reused across the client_portal-* apps

This is similar to the role currently managed by
client_portal_material_design, but should be framework agnostic so that
it can be used in all projects, regardless of whether they are using
angular/react/the-next-big-fucking-frontend-framework.

Note that it will, in fact, pull resources out of
client_portal_material_design, and be used to serve those assets to that
library.